Moved non-unit-test classes into net.lightbody.bmp.proxy.test.util class.
