+++ browsermob-core/src/test/java/net/lightbody/bmp/proxy/AddHeadersTest.java
@@ -1,5 +1,6 @@
 package net.lightbody.bmp.proxy;
 
+import net.lightbody.bmp.proxy.test.util.UnitTestServerTest;
 import net.lightbody.bmp.proxy.util.IOUtils;
 import org.apache.http.client.methods.HttpGet;
 import org.junit.Test;
@@ -8,7 +9,7 @@ import java.io.IOException;
 
 import static org.junit.Assert.assertTrue;
 
-public class AddHeadersTest extends DummyServerTest {
+public class AddHeadersTest extends UnitTestServerTest {
 
     @Test
     public void testAddHeadersToRequest() throws IOException {
