+++ browsermob-core/src/test/java/net/lightbody/bmp/proxy/CookieTest.java
@@ -3,6 +3,7 @@ package net.lightbody.bmp.proxy;
 import net.lightbody.bmp.core.har.Har;
 import net.lightbody.bmp.core.har.HarCookie;
 import net.lightbody.bmp.core.har.HarEntry;
+import net.lightbody.bmp.proxy.test.util.UnitTestServerTest;
 import net.lightbody.bmp.proxy.util.IOUtils;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.cookie.BasicClientCookie;
@@ -12,7 +13,7 @@ import org.junit.Test;
 
 import java.io.IOException;
 
-public class CookieTest extends DummyServerTest {
+public class CookieTest extends UnitTestServerTest {
     @Test
     public void testNoDoubleCookies() throws IOException {
         proxy.setCaptureContent(true);
