+++ browsermob-core/src/test/java/net/lightbody/bmp/proxy/MailingListIssuesTest.java
@@ -10,6 +10,7 @@ import net.lightbody.bmp.proxy.http.BrowserMobHttpRequest;
 import net.lightbody.bmp.proxy.http.BrowserMobHttpResponse;
 import net.lightbody.bmp.proxy.http.RequestInterceptor;
 import net.lightbody.bmp.proxy.http.ResponseInterceptor;
+import net.lightbody.bmp.proxy.test.util.UnitTestServerTest;
 import net.lightbody.bmp.proxy.util.IOUtils;
 import org.apache.http.HttpEntity;
 import org.apache.http.client.methods.HttpGet;
@@ -31,7 +32,7 @@ import java.net.URISyntaxException;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-public class MailingListIssuesTest extends DummyServerTest {
+public class MailingListIssuesTest extends UnitTestServerTest {
     @Test
     public void testThatInterceptorIsCalled() throws IOException, InterruptedException {
         final boolean[] interceptorHit = {false};
